+++ packages/bredon-types/src/builders/integer.js
@@ -1,18 +1,9 @@
 /* @flow */
 import type { IntegerNode } from '../../../../flowtypes/AST'
 
-export default function integer(
-  value: number,
-  isNegative?: boolean = false
-): IntegerNode {
-  const node: IntegerNode = {
+export default function integer(value: number): IntegerNode {
+  return {
     type: 'Integer',
     value
   }
-
-  if (isNegative) {
-    node.negative = true
-  }
-
-  return node
 }