+++ packages/bredon-types/src/builders/separator.js
@@ -1,7 +1,7 @@
 /* @flow */
 import type { SimpleNode } from '../../../../flowtypes/AST'
 
-export default function separator(kind: string): SimpleNode {
+export default function separator(kind: string = '/'): SimpleNode {
   return {
     type: 'Separator',
     value: kind