+++ packages/bredon-types/src/builders/stringLiteral.js
@@ -1,9 +1,12 @@
 /* @flow */
 import type { SimpleNode } from '../../../../flowtypes/AST'
 
-type Quote = 'double' | 'single'
+type Quote = '"' | "'"
 
-export default function stringLiteral(str: string, quote: Quote): SimpleNode {
+export default function stringLiteral(
+  str: string,
+  quote: Quote = "'"
+): SimpleNode {
   return {
     type: 'StringLiteral',
     value: str,