+++ packages/bredon/src/Generator.js
@@ -27,17 +27,15 @@ export default class Generator {
         return node.body.map(generateCSSValue).join('')
 
       case 'FunctionExpression':
-        return `${generateCSSValue(node.callee)}(${node.params
-          .map(generateCSSValue)
-          .join(',')})`
+        return `${node.callee}(${node.params.map(generateCSSValue).join(',')})`
 
       case 'Dimension':
-        return generateCSSValue(node.value) + node.unit
+        return node.value + node.unit
 
       case 'Float':
-        return `${node.integer.negative ? '-' : ''}${node.integer.value !== 0
-          ? node.integer.value
-          : ''}.${generateCSSValue(node.fractional)}`
+        return `${node.negative ? '-' : ''}${node.integer !== 0
+          ? node.integer
+          : ''}.${node.fractional}`
 
       case 'Operator':
         // for addition and substraction we use spacings left and right
@@ -49,15 +47,13 @@ export default class Generator {
       case 'StringLiteral':
         return node.quote + node.value + node.quote
 
-      case 'Integer':
-        return `${node.negative ? '-' : ''}${node.value}`
-
       case 'Identifier':
       case 'Important':
       case 'Keyword':
       case 'Parenthesis':
       case 'HexColor':
       case 'Separator':
+      case 'Integer':
       case 'URL':
         return node.value