+++ Changelog.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### 2.0.0
+* complete parser rewrite
+* better error messages
+* fixed many bugs
+* introducing `bredon-types` and `bredon-tools`
+
 ### 1.0.1
 * Whitespace inside `String` nodes now are preserved